step17 ログイン処理をhelperに切り出す
