+++ todo/spec/support/helpers/system_helper.rb
@@ -1,0 +1,8 @@
+module SystemHelper
+  def log_in_as(user)
+    visit login_path
+    fill_in 'session_name', with: user.name
+    fill_in 'session_password', with: user.password
+    click_on 'ログイン'
+  end
+end
